Math library - Splines
Build math library pkg and Add a new aerodynamic model, based on splines.

+++ last_letter_2/src/propulsion.cpp
@@ -14,9 +14,9 @@ void Propulsion::calcWrench()
 void Propulsion::calcAdditionalData()
 {
     // airspeed
-    float u_r = model->model_states.u - model->airdata.wind_x; 
-    float v_r = model->model_states.v - model->airdata.wind_y;
-    float w_r = model->model_states.w - model->airdata.wind_z;
+    double u_r = model->model_states.u - model->airdata.wind_x; 
+    double v_r = model->model_states.v - model->airdata.wind_y;
+    double w_r = model->model_states.w - model->airdata.wind_z;
     airspeed = sqrt(pow(u_r, 2) + pow(v_r, 2) + pow(w_r, 2));
 }
 
@@ -28,13 +28,13 @@ BeardEngine::BeardEngine(Model *parent) : Propulsion(parent)
 
 void BeardEngine::calcThrust()
 {
-    float delta_t = model->control_signals.delta_t;	
+    double delta_t = model->control_signals.delta_t;	
     prop_wrenches.thrust = 1.0 / 2.0 * rho * s_prop * c_prop * (pow(delta_t * k_motor, 2) - pow(airspeed, 2));
 }
 
 void BeardEngine::calcTorque()
 {
-    float delta_t = model->control_signals.delta_t;
+    double delta_t = model->control_signals.delta_t;
     prop_wrenches.torque = -k_t_p * pow((k_omega * delta_t), 2);
 }
